Various style and docs improvements
